Issue #46: Add ability to cancel an instance
`findPath()` now returns a unique number (even across multiple instances of easystar).  There's a new method `cancelPath()` which can be used to cancel a path calculation.

This can be used to avoid wasting effort, for instance if a game actor dies or changes its destination before the path is computed.