Add karma-benchmark, and some benchmarks.
There are two grids benchmarked:  One is maze-like, so the search will be deep but narrow.  The other is a mostly open field, so the search will be short but with more branches.

For both grids, the benchmarks are run three different ways: With no diagonals, diagonals but no corner-cutting, and diagonals with corner-cutting.