Cleanup in `calculate`
`isDoneCalculating` was unnecessary, the instance being looked at in `tilesToSearch[i].instance.isDoneCalculating` is always the current instance, and the point at which `isDoneCalculating` was being set to `true` was a fine place to go ahead and remove the instance.

`tilesToSearch` was unnecessary, it's sufficient to go straight to calling `checkAdjacentNode`.  This avoids constructing the `tilesToSearch` array and the tile objects in it.  That's not a big performance improvement, but it simplifies the code.