Add benchmarks (#48)
* Issue #46: Add ability to cancel an instance

`findPath()` now returns a unique number (even across multiple instances of easystar).  There's a new method `cancelPath()` which can be used to cancel a path calculation.

This can be used to avoid wasting effort, for instance if a game actor dies or changes its destination before the path is computed.

* Cleanup, and have cancelPath return boolean

* Missed a spot renaming _nextId.

Add a test case confirming that two instances have different IDs, which might have caught this mistake.

* Convert instances from array to map

Add instanceQueue to maintain the calculation order.  This change should make cancelling paths O(1) rather than O(n).

* update index.d.ts for path canelling changes.

* Add karma-benchmark, and some benchmarks.

There are two grids benchmarked:  One is maze-like, so the search will be deep but narrow.  The other is a mostly open field, so the search will be short but with more branches.

For both grids, the benchmarks are run three different ways: With no diagonals, diagonals but no corner-cutting, and diagonals with corner-cutting.

* Add karma-benchmark, and some benchmarks.

There are two grids benchmarked:  One is maze-like, so the search will be deep but narrow.  The other is a mostly open field, so the search will be short but with more branches.

For both grids, the benchmarks are run three different ways: With no diagonals, diagonals but no corner-cutting, and diagonals with corner-cutting.

* Bump version to 0.4.0

* Use synchronous benchmarks

Because these benchmarks use `enableAsync` there is no need to use the benchmarkjs `defer` flag.  This seems to avoid potential "Maximum call stack size exceeded" errors when the benchmarks are fast (possible bug in karma-benchmark).

* Add karma-benchmark, and some benchmarks.

There are two grids benchmarked:  One is maze-like, so the search will be deep but narrow.  The other is a mostly open field, so the search will be short but with more branches.

For both grids, the benchmarks are run three different ways: With no diagonals, diagonals but no corner-cutting, and diagonals with corner-cutting.

* Use synchronous benchmarks

Because these benchmarks use `enableAsync` there is no need to use the benchmarkjs `defer` flag.  This seems to avoid potential "Maximum call stack size exceeded" errors when the benchmarks are fast (possible bug in karma-benchmark).

* Add a 1000x1000 benchmark.

This pushes the limits of `karma-benchmarkjs-reporter`, in that it doesn't show precise results when ops take more than one second.

But a test like this is important to avoid optimizing the small-grid case at the expense of the large-grid case.